feat: Sprint14

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface UserService {
     UserDto createUser(UserDto userDto);
 
-    UserDto getUserById(Integer userId);
+    UserDto getUserById(Long userId);
 
     List<UserDto> getUsers();
 
-    UserDto updateUser(Integer userId, UserDto userDto);
+    UserDto updateUser(Long userId, UserDto userDto);
 
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 }
